+++ kernel/ipc/exo_disk.c
@@ -13,7 +13,7 @@
 
 [[nodiscard]] int exo_read_disk(struct exo_blockcap cap, void *dst,
                                 uint64_t off, uint64_t n) {
-  if (cap.owner != myproc()->pid || !cap_has_rights(cap.rights, EXO_RIGHT_R))
+  if (cap.owner != (uint32_t)myproc()->pid || !cap_has_rights(cap.rights, EXO_RIGHT_R))
     return -EPERM;
   struct buf b;
   uint64_t tot = 0;
@@ -45,7 +45,7 @@
 
 [[nodiscard]] int exo_write_disk(struct exo_blockcap cap, const void *src,
                                  uint64_t off, uint64_t n) {
-  if (cap.owner != myproc()->pid || !cap_has_rights(cap.rights, EXO_RIGHT_W))
+  if (cap.owner != (uint32_t)myproc()->pid || !cap_has_rights(cap.rights, EXO_RIGHT_W))
     return -EPERM;
   struct buf b;
   uint64_t tot = 0;
